Drag and drop works!
I got drag and drop to work.
It does have a bug where when you move the card into the droppable
area, that area will move or when there are no cards left the area
where the cards were at will disappear, anyone have any idea why this
happens? Its probably something about the css or the layout, if you
have an idea or want to find out why, go for it! It would be highly
appreciated!
Also don’t worry about the buttons I’m working on their implementations
So go ahead and check out the code!