Submit button works now
The game will check now if the cards that were just played are a valid
play, if they are the cards are moved to the left most position. This
is done to keep the board clean, and the cards are not draggable
anymore. Also the game will check if the card from the deck is used. I
fixed some game breaking bugs in this,  I’m sure more will arise as
more code is added but give it a try and run it.
Note: if the game is broken, or you break it, on chrome and safari you
can right click and press on inspect element to view the console and
see whats wrong. Write it down and tell me how you broke it so I can
take a look at it or you can! Doesn’t matter who finishes it as long as
this program is done.