Conquain Updated
The end button was updated to make cards that were put on the board and
not part of a valid play go back to the players hand
Card exchange is now implemented with the opponent class added.
I put a todo list at the bottom of the game with comments at where the
things will go in the game
Current ToDo List:
Increase the size of the board
Make the buttons disable, then enable it
Make a win screen/losing screen
Thanks!