Got the opponent to work
So now the computer will do something on its turn.
I took out the true from finding if you played a valid group of cards.
Also I added a transparent card that is not draggable so that the
container that holds the cards does not collapse.
I know I still need to change the opponents cards so you don’t see it
but for now its nice to see if the opponent object is working correctly.
I still need to put an image on the winning and loosing screen.