Fixed some bugs
So I fixed some bugs and now the cards for the computer are not shown.
The game has now been expanded to take the whole webpage.
If for some reason the game crashes please write down what the error is.
To do this right click anywhere on the page and inspect element. Click
on the console tab and write down or screenshot what the error is.
Thanks!